refactor(services): replace agentService with jobService in relevant components for improved modularity
