+++ web-app/src/lib/actions/job/action.ts
@@ -13,7 +13,7 @@ import {
   startJobInputSchema,
   StartJobInputSchemaType,
 } from "@/lib/schemas";
-import { agentService } from "@/lib/services";
+import { jobService } from "@/lib/services";
 import { Err, Ok, Result } from "@/lib/ts-res";
 
 export async function startJobWithInputData(
@@ -77,7 +77,7 @@ export async function startJobWithInputData(
       }
       const parsed = parsedResult.data;
 
-      const job = await agentService.startJob(parsed);
+      const job = await jobService.startJob(parsed);
 
       // Add success breadcrumb
       Sentry.addBreadcrumb({
@@ -260,7 +260,7 @@ export async function requestRefundJobByBlockchainIdentifier(
       });
     }
 
-    const job = await agentService.requestRefund(blockchainIdentifier);
+    const job = await jobService.requestRefund(blockchainIdentifier);
     return Ok({ job });
   } catch (error) {
     Sentry.withScope((scope) => {
